Introduce "validations table" feature; update API
This introduces the concept of a validations table that contains a list of
validation functions that are applied to overall AI data during data input.

Validation functions may be enabled or disabled, unless they are defined as
"locked".

The validation is achieved by iterating over the table and calling the
validation function pointers for each enabled entry.